# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CreateContactSwagger:
    properties:
      answer_email:
        example: ""
        type: string
      contact_email:
        example: ""
        type: string
      facebook:
        example: ""
        type: string
      linkedin:
        example: ""
        type: string
      phone:
        example: ""
        type: string
      telegram:
        example: ""
        type: string
    type: object
...
  model.CreateHeroSwagger:
    properties:
      description:
        example: some description
        type: string
      image:
        example: stunning image
        type: string
      title:
        example: some title
        type: string
    type: object
...
  model.Response:
    properties:
      code:
        type: integer
      messsage:
        type: string
    type: object
...
  model.UpdateContactSwagger:
    properties:
      answer_email:
        example: ""
        type: string
      contact_email:
        example: ""
        type: string
      facebook:
        example: ""
        type: string
      id:
        example: ""
        type: string
      linkedin:
        example: ""
        type: string
      phone:
        example: ""
        type: string
      telegram:
        example: ""
        type: string
    type: object
  model.UpdateHeroSwagger:
    properties:
      description:
        example: some description
        type: string
      id:
        example: ""
        type: string
      image:
        example: stunning image
        type: string
      title:
        example: some title
        type: string
    type: object
...
host: walking-school.site
info:
  contact:
    email: etverya12@gmail.com
    name: Yehor Tverytinov
  title: Walking-School backend API
  version: "1.0"
paths:
  /contact:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all contacts.
      tags:
      - contact
    post:
      consumes:
      - application/json
      parameters:
      - description: Contact
        in: body
        name: Contact
        required: true
        schema:
          $ref: '#/definitions/model.CreateContactSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create contact .
      tags:
      - contact
    put:
      consumes:
      - application/json
      parameters:
      - description: Contact
        in: body
        name: Contact
        required: true
        schema:
          $ref: '#/definitions/model.UpdateContactSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update contact by id.
      tags:
      - contact
...
  /hero:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all heros.
      tags:
      - hero
    post:
      consumes:
      - application/json
      parameters:
      - description: Hero
        in: body
        name: Hero
        required: true
        schema:
          $ref: '#/definitions/model.CreateHeroSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create hero.
      tags:
      - hero
    put:
      consumes:
      - application/json
      parameters:
      - description: Hero
        in: body
        name: Hero
        required: true
        schema:
          $ref: '#/definitions/model.UpdateHeroSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update hero by id.
      tags:
      - hero
  /hero/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete hero by id.
      tags:
      - hero
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get hero by id.
      tags:
      - hero
...
swagger: "2.0"